moved all forecaster examples and benchmarks into own file
