+++ forecaster_test.go
@@ -3,12 +3,9 @@ package forecaster
 import (
 	"fmt"
 	"math"
-	"os"
-	"runtime/debug"
 	"testing"
 	"time"
 
-	"github.com/goccy/go-json"
 	"gonum.org/v1/gonum/mat"
 
 	"github.com/aouyang1/go-forecaster/feature"
@@ -16,7 +13,6 @@ import (
 	"github.com/aouyang1/go-forecaster/forecast/options"
 	"github.com/aouyang1/go-forecaster/forecast/util"
 	"github.com/aouyang1/go-forecaster/timedataset"
-	"github.com/pkg/profile"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -825,436 +821,6 @@ func TestForecaster(t *testing.T) {
 	}
 }
 
-func generateExampleSeries() ([]time.Time, []float64) {
-	// create a daily sine wave at minutely with one week
-	minutes := 28 * 24 * 60
-	t := timedataset.GenerateT(minutes, time.Minute, time.Now)
-	y := make(timedataset.Series, minutes)
-
-	period := 86400.0
-	y.Add(timedataset.GenerateConstY(minutes, 98.3)).
-		Add(timedataset.GenerateWaveY(t, 10.5, period, 1.0, 2*60*60)).
-		Add(timedataset.GenerateWaveY(t, 10.5, period, 3.0, 2.0*60*60+period/2.0/2.0/3.0)).
-		Add(timedataset.GenerateWaveY(t, 23.4, period, 7.0, 6.0*60*60+period/2.0/2.0/3.0).MaskWithTimeRange(t[minutes*4/16], t[minutes*5/16], t)).
-		Add(timedataset.GenerateWaveY(t, -7.3, period, 3.0, 2*60*60+period/2.0/2.0/3.0).MaskWithWeekend(t)).
-		Add(timedataset.GenerateNoise(t, 3.2, 3.2, period, 5.0, 0.0)).
-		Add(timedataset.GenerateChange(t, t[minutes/2], 10.0, 0.0)).
-		Add(timedataset.GenerateChange(t, t[minutes*19/20], -70.0, 0.0)).
-		SetConst(t, 2.7, t[minutes/3], t[minutes/3+minutes/20]).
-		SetConst(t, 175.7, t[minutes*2/3], t[minutes*2/3+minutes/80])
-
-	return t, y
-}
-
-func runForecastExample(opt *Options, t []time.Time, y []float64, filename string) error {
-	f, err := New(opt)
-	if err != nil {
-		return err
-	}
-	if err := f.Fit(t, y); err != nil {
-		return err
-	}
-
-	m, err := f.Model()
-	if err != nil {
-		return err
-	}
-	if err := m.TablePrint(os.Stderr); err != nil {
-		return err
-	}
-
-	file, err := os.Create(filename)
-	if err != nil {
-		return err
-	}
-
-	return f.PlotFit(file, nil)
-}
-
-func recoverForecastPanic(t *testing.T) {
-	if r := recover(); r != nil {
-		if t != nil {
-			t.Errorf("panic: %v\n", r)
-		} else {
-			fmt.Printf("panic: %v\n", r)
-		}
-		debug.PrintStack()
-	}
-}
-
-func setupWithOutliers() ([]time.Time, []float64, *Options) {
-	t, y := generateExampleSeries()
-
-	changepoints := []options.Changepoint{
-		options.NewChangepoint("anomaly1", t[len(t)/2]),
-		options.NewChangepoint("anomaly2", t[len(t)*19/20]),
-	}
-	events := []options.Event{
-		options.NewEvent("custom_event", t[len(t)*4/16], t[len(t)*5/16]),
-	}
-
-	regularization := []float64{0.0, 1.0, 10.0, 100.0, 1000.0, 10000.0}
-	opt := &Options{
-		SeriesOptions: &SeriesOptions{
-			ForecastOptions: &options.Options{
-				Regularization: regularization,
-				SeasonalityOptions: options.SeasonalityOptions{
-					SeasonalityConfigs: []options.SeasonalityConfig{
-						options.NewDailySeasonalityConfig(12),
-						options.NewWeeklySeasonalityConfig(12),
-					},
-				},
-				Iterations: 500,
-				Tolerance:  1e-3,
-				ChangepointOptions: options.ChangepointOptions{
-					Changepoints: changepoints,
-				},
-				WeekendOptions: options.WeekendOptions{
-					Enabled: true,
-				},
-				EventOptions: options.EventOptions{
-					Events: events,
-				},
-				DSTOptions: options.DSTOptions{
-					Enabled: false,
-				},
-			},
-			OutlierOptions: &OutlierOptions{
-				NumPasses:       3,
-				TukeyFactor:     1.5,
-				LowerPercentile: 0.25,
-				UpperPercentile: 0.75,
-			},
-		},
-		UncertaintyOptions: &UncertaintyOptions{
-			ForecastOptions: &options.Options{
-				Regularization: regularization,
-				SeasonalityOptions: options.SeasonalityOptions{
-					SeasonalityConfigs: []options.SeasonalityConfig{
-						options.NewDailySeasonalityConfig(12),
-						options.NewWeeklySeasonalityConfig(12),
-					},
-				},
-				Iterations: 250,
-				Tolerance:  1e-2,
-				ChangepointOptions: options.ChangepointOptions{
-					Changepoints: nil,
-				},
-				WeekendOptions: options.WeekendOptions{
-					Enabled: true,
-				},
-				EventOptions: options.EventOptions{
-					Events: events,
-				},
-				DSTOptions: options.DSTOptions{
-					Enabled: false,
-				},
-			},
-			ResidualWindow: 100,
-			ResidualZscore: 4.0,
-		},
-		MinUncertaintyValue: 13.0,
-	}
-
-	return t, y, opt
-}
-
-func Example_forecasterWithOutliers() {
-	t, y, opt := setupWithOutliers()
-
-	defer recoverForecastPanic(nil)
-
-	if err := runForecastExample(opt, t, y, "examples/forecaster.html"); err != nil {
-		panic(err)
-	}
-	// Output:
-}
-
-func Example_forecasterWithOutliersConstantUncertainty() {
-	t, y, opt := setupWithOutliers()
-	opt.UncertaintyOptions.Percentage = 0.10
-	opt.MinUncertaintyValue = 0.0
-
-	defer recoverForecastPanic(nil)
-
-	if err := runForecastExample(opt, t, y, "examples/forecaster_constant_uncertainty.html"); err != nil {
-		panic(err)
-	}
-	// Output:
-}
-
-func generateExampleSeriesWithTrend() ([]time.Time, []float64) {
-	// create a daily sine wave at minutely with one week
-	minutes := 4 * 24 * 60
-	t := timedataset.GenerateT(minutes, time.Minute, time.Now)
-	y := make(timedataset.Series, minutes)
-
-	period := 86400.0
-	y.Add(timedataset.GenerateConstY(minutes, 98.3)).
-		Add(timedataset.GenerateWaveY(t, 10.5, period, 1.0, 2*60*60)).
-		Add(timedataset.GenerateWaveY(t, 10.5, period, 3.0, 2.0*60*60+period/2.0/2.0/3.0)).
-		Add(timedataset.GenerateNoise(t, 3.2, 3.2, period, 5.0, 0.0)).
-		Add(timedataset.GenerateChange(t, t[minutes/2], 0.0, 40.0/(t[minutes*17/20].Sub(t[minutes/2]).Minutes()))).
-		Add(timedataset.GenerateChange(t, t[minutes*17/20], -40.0, -40.0/(t[minutes*17/20].Sub(t[minutes/2]).Minutes())))
-
-	return t, y
-}
-
-func generateExampleSeriesWithGlobalTrend() ([]time.Time, []float64) {
-	// create a daily sine wave at minutely with one week
-	minutes := 14 * 24 * 60
-	t := timedataset.GenerateT(minutes, time.Minute, time.Now)
-	y := make(timedataset.Series, minutes)
-
-	period := 86400.0
-	y.Add(timedataset.GenerateConstY(minutes, 98.3)).
-		Add(timedataset.GenerateWaveY(t, 10.5, period, 1.0, 2*60*60)).
-		Add(timedataset.GenerateWaveY(t, 10.5, period, 3.0, 2.0*60*60+period/2.0/2.0/3.0)).
-		Add(timedataset.GenerateNoise(t, 3.2, 3.2, period, 5.0, 0.0)).
-		Add(timedataset.GenerateChange(t, t[0], 0.0, 40.0/(t[minutes*17/20].Sub(t[minutes/2]).Minutes())))
-
-	return t, y
-}
-
-/*
-func Example_forecasterAutoChangepoint() {
-	t, y := generateExampleSeries()
-
-	regularization := []float64{0.0, 1.0, 10.0, 100.0, 1000.0, 10000.0}
-	opt := &Options{
-		SeriesOptions: &SeriesOptions{
-			ForecastOptions: &options.Options{
-				Regularization: regularization,
-				SeasonalityOptions: options.SeasonalityOptions{
-					SeasonalityConfigs: []options.SeasonalityConfig{
-						options.NewDailySeasonalityConfig(12),
-						options.NewWeeklySeasonalityConfig(12),
-					},
-				},
-				Iterations: 500,
-				Tolerance:  1e-3,
-				ChangepointOptions: options.ChangepointOptions{
-					Auto:                true,
-					AutoNumChangepoints: 100,
-					EnableGrowth:        false,
-				},
-			},
-			OutlierOptions: NewOutlierOptions(),
-		},
-		UncertaintyOptions: &UncertaintyOptions{
-			ForecastOptions: &options.Options{
-				Regularization: regularization,
-				SeasonalityOptions: options.SeasonalityOptions{
-					SeasonalityConfigs: []options.SeasonalityConfig{
-						options.NewDailySeasonalityConfig(12),
-						options.NewWeeklySeasonalityConfig(12),
-					},
-				},
-				Iterations: 250,
-				Tolerance:  1e-2,
-				ChangepointOptions: options.ChangepointOptions{
-					Auto:         false,
-					Changepoints: []options.Changepoint{},
-				},
-			},
-			ResidualWindow: 100,
-			ResidualZscore: 4.0,
-		},
-	}
-	opt.SetMinValue(0.0)
-	opt.SetMaxValue(170.0)
-
-	defer recoverForecastPanic(nil)
-
-	if err := runForecastExample(opt, t, y, "examples/forecaster_auto_changepoint.html"); err != nil {
-		panic(err)
-	}
-	// Output:
-}
-*/
-
-func Example_forecasterWithTrend() {
-	t, y := generateExampleSeriesWithTrend()
-
-	changepoints := []options.Changepoint{
-		options.NewChangepoint("trendstart", t[len(t)/2]),
-		options.NewChangepoint("rebaseline", t[len(t)*17/20]),
-	}
-
-	opt := &Options{
-		SeriesOptions: &SeriesOptions{
-			ForecastOptions: &options.Options{
-				SeasonalityOptions: options.SeasonalityOptions{
-					SeasonalityConfigs: []options.SeasonalityConfig{
-						options.NewDailySeasonalityConfig(12),
-						options.NewWeeklySeasonalityConfig(12),
-					},
-				},
-				Iterations: 500,
-				Tolerance:  1e-3,
-				ChangepointOptions: options.ChangepointOptions{
-					Changepoints: changepoints,
-					EnableGrowth: true,
-				},
-			},
-			OutlierOptions: NewOutlierOptions(),
-		},
-		UncertaintyOptions: &UncertaintyOptions{
-			ForecastOptions: &options.Options{
-				SeasonalityOptions: options.SeasonalityOptions{
-					SeasonalityConfigs: []options.SeasonalityConfig{
-						options.NewDailySeasonalityConfig(12),
-						options.NewWeeklySeasonalityConfig(12),
-					},
-				},
-				Iterations: 250,
-				Tolerance:  1e-2,
-				ChangepointOptions: options.ChangepointOptions{
-					Changepoints: nil,
-				},
-			},
-			ResidualWindow: 100,
-			ResidualZscore: 4.0,
-		},
-	}
-
-	defer recoverForecastPanic(nil)
-
-	if err := runForecastExample(opt, t, y, "examples/forecaster_with_trend.html"); err != nil {
-		panic(err)
-	}
-	// Output:
-}
-
-func Example_forecasterWithGlobalTrend() {
-	t, y := generateExampleSeriesWithGlobalTrend()
-
-	opt := &Options{
-		SeriesOptions: &SeriesOptions{
-			ForecastOptions: &options.Options{
-				SeasonalityOptions: options.SeasonalityOptions{
-					SeasonalityConfigs: []options.SeasonalityConfig{
-						options.NewDailySeasonalityConfig(12),
-						options.NewWeeklySeasonalityConfig(12),
-					},
-				},
-				Iterations: 500,
-				Tolerance:  1e-3,
-				GrowthType: feature.GrowthLinear,
-			},
-			OutlierOptions: NewOutlierOptions(),
-		},
-		UncertaintyOptions: &UncertaintyOptions{
-			ForecastOptions: &options.Options{
-				SeasonalityOptions: options.SeasonalityOptions{
-					SeasonalityConfigs: []options.SeasonalityConfig{
-						options.NewDailySeasonalityConfig(12),
-						options.NewWeeklySeasonalityConfig(12),
-					},
-				},
-				Iterations: 250,
-				Tolerance:  1e-2,
-				ChangepointOptions: options.ChangepointOptions{
-					Changepoints: nil,
-				},
-			},
-			ResidualWindow: 100,
-			ResidualZscore: 4.0,
-		},
-	}
-
-	defer recoverForecastPanic(nil)
-
-	if err := runForecastExample(opt, t, y, "examples/forecaster_with_global_trend.html"); err != nil {
-		panic(err)
-	}
-	// Output:
-}
-
-func generateExamplePulseSeries() ([]time.Time, []float64) {
-	// create a daily sine wave at minutely with one week
-	minutes := 28 * 24 * 60
-	t := timedataset.GenerateT(minutes, time.Minute, time.Now)
-	y := make(timedataset.Series, minutes)
-
-	period := (3 * time.Hour).Seconds()
-	y.Add(timedataset.GenerateConstY(len(t), 30)).
-		Add(timedataset.GeneratePulseY(t, 100.0, period, 1.0, 0.0, 0.05)).
-		Add(timedataset.GenerateNoise(t, 3.2, 0.0, period, 5.0, 0.0))
-
-	return t, y
-}
-
-func setupWithPulses() ([]time.Time, []float64, *Options) {
-	t, y := generateExamplePulseSeries()
-
-	changepoints := []options.Changepoint{}
-	events := []options.Event{}
-
-	regularization := []float64{0.0, 1.0, 10.0, 100.0, 1000.0, 10000.0}
-	opt := &Options{
-		SeriesOptions: &SeriesOptions{
-			ForecastOptions: &options.Options{
-				Regularization: regularization,
-				SeasonalityOptions: options.SeasonalityOptions{
-					SeasonalityConfigs: []options.SeasonalityConfig{
-						options.NewSeasonalityConfig("pulse", 3*time.Hour, 24),
-					},
-				},
-				Iterations: 500,
-				Tolerance:  1e-3,
-				ChangepointOptions: options.ChangepointOptions{
-					Changepoints: changepoints,
-				},
-				WeekendOptions: options.WeekendOptions{
-					Enabled: false,
-				},
-				EventOptions: options.EventOptions{
-					Events: events,
-				},
-			},
-			OutlierOptions: NewOutlierOptions(),
-		},
-		UncertaintyOptions: &UncertaintyOptions{
-			ForecastOptions: &options.Options{
-				Regularization: regularization,
-				SeasonalityOptions: options.SeasonalityOptions{
-					SeasonalityConfigs: []options.SeasonalityConfig{
-						options.NewDailySeasonalityConfig(12),
-						options.NewWeeklySeasonalityConfig(12),
-					},
-				},
-				Iterations: 250,
-				Tolerance:  1e-2,
-				ChangepointOptions: options.ChangepointOptions{
-					Changepoints: nil,
-				},
-				WeekendOptions: options.WeekendOptions{
-					Enabled: false,
-				},
-				EventOptions: options.EventOptions{
-					Events: events,
-				},
-			},
-			ResidualWindow: 100,
-			ResidualZscore: 4.0,
-		},
-	}
-
-	return t, y, opt
-}
-
-func Example_forecasterWithPulses() {
-	t, y, opt := setupWithPulses()
-
-	defer recoverForecastPanic(nil)
-
-	if err := runForecastExample(opt, t, y, "examples/forecaster_pulses.html"); err != nil {
-		panic(err)
-	}
-	// Output:
-}
-
 func TestFixedPercentageUncertainty(t *testing.T) {
 	nowFunc := func() time.Time {
 		return time.Date(1970, 2, 1, 0, 0, 0, 0, time.UTC)
@@ -1336,69 +902,4 @@ func TestMatrixMulWithNaN(t *testing.T) {
 	fc := mat.Formatted(&c, mat.Prefix("    "), mat.Squeeze())
 	fmt.Printf("c = %v", fc)
 	// Output: c = [16  0  12  NaN]
-}
-
-var benchPredictRes *Results
-
-func BenchmarkTrainToModel(b *testing.B) {
-	t, y, opt := setupWithOutliers()
-
-	var f *Forecaster
-	var err error
-
-	b.ResetTimer()
-	for b.Loop() {
-		f, err = New(opt)
-		if err != nil {
-			panic(err)
-		}
-
-		if err := f.Fit(t, y); err != nil {
-			panic(err)
-		}
-	}
-
-	m, err := f.Model()
-	if err != nil {
-		panic(err)
-	}
-
-	bytes, err := json.MarshalIndent(m, "", "  ")
-	if err != nil {
-		panic(err)
-	}
-
-	if err := os.WriteFile("benchmark_model.json", bytes, 0o644); err != nil {
-		panic(err)
-	}
-}
-
-func BenchmarkPredictFromModel(b *testing.B) {
-	bytes, err := os.ReadFile("benchmark_model.json")
-	if err != nil {
-		panic(err)
-	}
-
-	var model Model
-	if err := json.Unmarshal(bytes, &model); err != nil {
-		panic(err)
-	}
-	f, err := NewFromModel(model)
-	if err != nil {
-		panic(err)
-	}
-
-	input := make([]time.Time, 0, 2)
-	ct := time.Now()
-	for i := range cap(input) {
-		input = append(input, ct.Add(time.Duration(i)*time.Minute))
-	}
-	b.ResetTimer()
-	defer profile.Start(profile.CPUProfile, profile.ProfilePath(".")).Stop()
-	for b.Loop() {
-		benchPredictRes, err = f.Predict(input)
-		if err != nil {
-			panic(err)
-		}
-	}
 }
